add weather support update packages for refactor
