+++ pom.xml
@@ -39,12 +39,12 @@
     <dependency>
       <groupId>com.bigboxer23</groupId>
       <artifactId>solar-moon-common</artifactId>
-      <version>1.5.34</version>
+      <version>1.6.4</version>
     </dependency>
     <dependency>
       <groupId>com.bigboxer23</groupId>
       <artifactId>solar-moon-common</artifactId>
-      <version>1.5.34</version>
+      <version>1.6.4</version>
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
