+++ src/main/java/com/bigboxer23/solar_moon/alarm/ExtendedAlarmComponent.java
@@ -1,11 +1,9 @@
 package com.bigboxer23.solar_moon.alarm;
 
-import com.bigboxer23.solar_moon.AlarmComponent;
-import com.bigboxer23.solar_moon.DeviceComponent;
-import com.bigboxer23.solar_moon.OpenWeatherComponent;
 import com.bigboxer23.solar_moon.data.Alarm;
+import com.bigboxer23.solar_moon.device.DeviceComponent;
 import com.bigboxer23.solar_moon.notifications.NotificationComponent;
-import com.bigboxer23.solar_moon.open_search.OpenSearchComponent;
+import com.bigboxer23.solar_moon.search.OpenSearchComponent;
 import com.bigboxer23.solar_moon.util.TableCreationUtils;
 import java.util.Arrays;
 import org.springframework.stereotype.Component;
@@ -15,11 +13,10 @@ import org.springframework.stereotype.Component;
 public class ExtendedAlarmComponent extends AlarmComponent {
 
 	public ExtendedAlarmComponent(
-			OpenWeatherComponent openWeatherComponent,
 			DeviceComponent deviceComponent,
 			OpenSearchComponent OSComponent,
 			NotificationComponent notificationComponent) {
-		super(openWeatherComponent, deviceComponent, OSComponent, notificationComponent);
+		super(deviceComponent, OSComponent, notificationComponent);
 	}
 
 	public void createAlarmTable() {
