+++ src/main/java/com/bigboxer23/solar_moon/device/ExtendedDeviceUpdateComponent.java
@@ -1,6 +1,5 @@
 package com.bigboxer23.solar_moon.device;
 
-import com.bigboxer23.solar_moon.DeviceUpdateComponent;
 import com.bigboxer23.solar_moon.data.DeviceUpdateData;
 import com.bigboxer23.solar_moon.util.TableCreationUtils;
 import java.util.Collections;
