+++ src/main/java/com/bigboxer23/solar_moon/search/ExtendedOpenSearchComponent.java
@@ -1,6 +1,5 @@
 package com.bigboxer23.solar_moon.search;
 
-import com.bigboxer23.solar_moon.open_search.OpenSearchComponent;
 import org.springframework.stereotype.Component;
 
 /** */
