+++ src/main/java/com/bigboxer23/solar_moon/subscription/ExtendedSubscriptionComponent.java
@@ -1,6 +1,5 @@
 package com.bigboxer23.solar_moon.subscription;
 
-import com.bigboxer23.solar_moon.SubscriptionComponent;
 import com.bigboxer23.solar_moon.util.TableCreationUtils;
 import java.util.Collections;
 import org.springframework.stereotype.Component;
