+++ src/main/java/com/bigboxer23/solar_moon/util/AuthUtil.java
@@ -1,6 +1,6 @@
 package com.bigboxer23.solar_moon.util;
 
-import com.bigboxer23.solar_moon.CustomerComponent;
+import com.bigboxer23.solar_moon.customer.CustomerComponent;
 import com.bigboxer23.solar_moon.data.Customer;
 import com.bigboxer23.solar_moon.lambda.data.CognitoUserAttributes;
 import com.squareup.moshi.Moshi;
