+++ src/main/java/com/bigboxer23/solar_moon/weather/ExtendedOpenWeatherComponent.java
@@ -1,6 +1,5 @@
 package com.bigboxer23.solar_moon.weather;
 
-import com.bigboxer23.solar_moon.OpenWeatherComponent;
 import org.springframework.stereotype.Component;
 
 /** */
