+++ src/test/java/com/bigboxer23/solar_moon/weather/TestExtendedPirateWeatherComponent.java
@@ -1,0 +1,14 @@
+package com.bigboxer23.solar_moon.weather;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ *
+ */
+public class TestExtendedPirateWeatherComponent
+{
+	@Test
+	public void createTable() {
+		new ExtendedPirateWeatherComponent().createWeatherTable();
+	}
+}
